fix: crates.io API 403 and download-count overflow
crates.io rejects requests without a User-Agent header (403); add
one via an OkHttp interceptor. Also widen downloads/recent_downloads
fields from Int to Long — the live API returns crate download counts
that overflow Int (e.g. 2,222,988,134), which crashed Moshi parsing
with JsonDataException and left the summary screen stuck on a bare
"Error" state. Found by running the app end-to-end against the real
API. Also logs the underlying exception in CratesSummaryViewModel so
future failures are diagnosable from logcat.

+++ app/src/main/kotlin/io/github/andraantariksa/crates/feature_crates/data/source/remote/model/common/Crate.kt
@@ -6,7 +6,7 @@ data class Crate(
         val created_at: String,
         val description: String,
         val documentation: String?,
-        val downloads: Int,
+        val downloads: Long,
         val exact_match: Boolean,
         val homepage: String,
         val id: String,
@@ -16,7 +16,7 @@ data class Crate(
         val max_version: String,
         val name: String,
         val newest_version: String,
-        val recent_downloads: Int,
+        val recent_downloads: Long,
         val repository: String,
         val updated_at: String,
         val versions: List<Int>
